pyproject.toml
[tool.poetry]
name = "pong"
version = "0.1.0"
description = ""
authors = ["Aliphinos-Fylakas <pessoal.carlosveras@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pygame-ce = "^2.1.3"

[tool.poetry.group.test.dependencies]
pytest = "^7.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
